[Refactor] 아이템 관련, 인벤토리  클래스 namespace 제거, 폴더 이동
