--- Assets/SHG/Scripts/Items/ICraftable.cs
@@ -1,11 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-namespace SHG
-{
-  public class ICraftable 
-  {
-
-  }
-}
