--- Assets/SHG/Scripts/Items/IUsable.cs
@@ -1,9 +1,0 @@
-using System;
-
-namespace SHG
-{
-  public interface IUsable
-  {
-    public Action<IUsable> OnUsed { get; set; }
-  }
-}
