--- Assets/SHG/Scripts/Items/Item.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e6adee6fb8bac7176976c2fc06cea7cf
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
