--- Assets/SHG/Scripts/Items/ItemData.cs
@@ -1,28 +1,0 @@
-using System;
-using UnityEngine;
-using EditorAttributes;
-
-namespace SHG
-{
-  [Serializable]
-  public abstract class ItemData : ScriptableObject
-  {
-    [HideInInspector]
-    public string Name => this.itemName;
-    [HideInInspector]
-    public Sprite Image => this.image;
-    [HideInInspector]
-    public GameObject Prefab => this.prefab;
-
-    [SerializeField, Validate("Empty item name", nameof(IsNameEmpty), MessageMode.Error, buildKiller: true)]
-    protected string itemName;
-    [SerializeField, AssetPreview(64f, 64f), Validate("Image is none", nameof(IsImageNone), MessageMode.Warning)]
-    protected Sprite image;
-    [SerializeField, AssetPreview(64f, 64f), Validate("Prefab is none", nameof(IsPrefabNone), MessageMode.Warning)]
-    protected GameObject prefab;
-
-    protected bool IsNameEmpty() => this.itemName == null || this.itemName.Replace(" ", "").Length == 0;
-    protected bool IsImageNone() => this.Image == null;
-    protected bool IsPrefabNone() => this.Prefab == null;
-  }
-}
