+++ Assets/SHG/Scripts/Temp.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 58adcd8f38db436f09f4daeeef50db44
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
