+++ Assets/Scripts/Intefaces.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d199cba297edc6745a3cd8f83bbd440f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
