+++ Assets/Scripts/Intefaces/IObservableObject.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: fcc4062a9ce0be4d2b06650ea5fb6abe
+guid: 7d0cefe376e404f1186313ef7f3fe782
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
