+++ Assets/Scripts/Intefaces/IRecoveryable.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 2f75c9f89c7970284bc240532b3c910a
+guid: 7567e40ef4f206dfc99809a82f23ea55
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
