+++ Assets/Scripts/Intefaces/IUsable.cs
@@ -1,0 +1,6 @@
+using System;
+
+public interface IUsable
+{
+  public Action<IUsable> OnUsed { get; set; }
+}
