+++ Assets/Scripts/Intefaces/IUsable.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3e34c8d2850a9dc1989c54a19e19962a
+guid: eda01c6ce1f572a3fa9e1f431e47ce01
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
