+++ Assets/Scripts/Inventory/Inventory.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5e17e772bb742b2bda42aeb81ced6e9a
+guid: 53bcc406b00ae2de080ee72ca6ffd7f5
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
