+++ Assets/Scripts/Items.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3d28e695d0f58d1c69bd194f1cce7a44
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
