+++ Assets/Scripts/Items/Item.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: fac7b084284749ad8b6f46a0cc675632
+guid: c9682f9689c91da6fb7b6e016009c5b6
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
