+++ Assets/Scripts/Items/ItemData.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 1f7bc2599ff9c1507bb39dec3a21ac95
+guid: 79d37669f9644b079b85ec556ee2d9ad
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
