+++ Assets/Scripts/Items/ItemHolder.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3f0848c2e0a9e8d5dab59a9b94b30bf7
+guid: 7ab39b5713b0be7a5a685c085426c13b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
