+++ Assets/Scripts/Items/RecoveryItem.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: abb763c9f33aac7318eb3ed7a3994bca
+guid: 28314cf08c32980f29365dc574c2b8c0
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
