+++ Assets/Scripts/Items/RecoveryItemData.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: eac473a93e244eeddb2e55e3f3ce7844
+guid: 7567cf3238682da98bdaeed661d52583
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
