+++ Assets/Scripts/Patterns/ObservableValue.cs
