+++ Assets/Scripts/Patterns/ObservableValue.cs.meta
